Improve the English Translation of Solemnity Titles
Several solemnities had somewhat strange phrasing for their English
titles. E.g. "Octave Day of Christmas, of Mary, Mother of God". This
commit improves the titles to more closely match the text used in the
English translation of the liturgy of the hours. In some cases, it is
impossible to match the formatting (since we are limited to plain text),
so we make minor changes where appropriate.

This fixes #26.