Improve developer instructions in README
When I started working on a few minor contributions last month, it was
not obvious to me how it should be run. I tried a couple things:

    $ ./bin/calendariumrom
    Traceback (most recent call last):
    	2: from ./bin/calendariumrom:3:in `<main>'
    	1: from /usr/lib/ruby/2.5.0/rubygems/core_ext/kernel_require.rb:59:in `require'
    /usr/lib/ruby/2.5.0/rubygems/core_ext/kernel_require.rb:59:in `require': cannot load such file -- calendarium-romanum (LoadError)

    $ bundle exec bin/calendariumrom
    bundler: failed to load command: bin/calendariumrom (bin/calendariumrom)
    LoadError: cannot load such file -- calendarium-romanum
      bin/calendariumrom:3:in `require'
      bin/calendariumrom:3:in `<top (required)>'

Unfortunately, these all resulted in errors. The solution is to run with
`-Ilib`:

    $ bundle exec ruby -Ilib bin/calendariumrom
    Commands:
      calendariumrom calendars          # lists calendars available for querying

This commit adds some instructions to the README so developers know how
to do this if they want to run calendariumrom from the source code.