Sanctorale#update allows celebrations changing date
Since the introduction of Celebration symbol uniqueness check it was
impossible to load the calendar of the diocese of Litoměřice
(which moves one memorial from the General Roman Calendar to an earlier
day due to a collision with the Dedication of the cathedral),
I only stumbled upon it now.

Unfortunately this change breaks strong data consistency -
the check for data uniqueness can only be done at the end of #update
and if it fails, the Sanctorale instance is left in an inconsistent
state (having multiple Celebrations with the same symbol).
I don't think there is an efficient way to fix this.