Easter Triduum as separate liturgical season
"Normae universales" 18-21 set the Triduum clearly apart from Lent and
Eastertide. The modelling of liturgical seasons used so far was
inaccurate.

This is a major breaking change, but given we are still in 0.x and given
it's importance in respect of implementing the liturgical law correctly,
it will be included in the next minor version.