Transfers: on collision of two solemnities of equal rank Temporale wins
... and the Sanctorale solemnity is transferred.

So far the behaviour in such cases was unspecified (not covered by
specs) and in fact the Temporale solemnity was being selected for
transfer, due to the way how Enumerable#sort_by behaves when comparing
two equal values (it preserves the original order).

The rule that in cases of colliding solemnities of equal rank the one
provided by the Temporale cycle must win is not explicitly pronounced in
the liturgical law, possibly because the liturgical law operates with a
set of celebration ranks with higher granularity than
calendarium-romanum (cf. GNLYC 59).
But the rule is in accord with general architecture of the reformed
Roman liturgical calendar (cf. GNLYC 50 a; SC 108)
and results of it's application mostly (if not always) match those
that would be produced if calendarium-romanum used a rank system
allocating a separate rank for every single subitem mentioned in the
Table of Liturgical Days.

Introduction of this rule also accidentally fixes the duplication and
loss of conflicting solemnities of equal rank (#80), but the fix is
not reliable, as it only works well for the current combination of
Transfers logic and celebration picking logic in
Calendar#celebrations_for.