failing spec proving current implementation of transfers unreliable
The default implementation of solemnity transfers (now) works well only
due to a combination of compatible logic in Transfers and
Calendar#celebrations_for, both preferring temporale solemnities over
sanctorale ones if both meet on the same date with the same rank.
But if this consent between Calendar and Transfers is broken, it stops
working and the solemnity loss and duplication bug (#80) is back.